header bugfix
